Addes missing cast to int for java array

+++ ui/org.omnetpp.common.core/src/org/omnetpp/common/color/ColorFactory.java
@@ -854,7 +854,7 @@ public class ColorFactory {
      * Returns a "good" light color.
      */
     public static Color getGoodLightColor(long i) {
-        return goodLightColors[i % goodLightColors.length];
+        return goodLightColors[(int) (i % goodLightColors.length)];
     }
 
     public static Color[] getGoodDarkColors() {
